:truck: Restructured certificate routing and file structure..
