+++ app/__init__.py
@@ -6,7 +6,6 @@ from flask_mail import Mail
 from flask_bootstrap import Bootstrap
 from flask_apscheduler import APScheduler
 from dotenv import load_dotenv
-import pdfkit
 
 load_dotenv()
 
@@ -49,8 +48,6 @@ login.login_view = "auth.login"
 login.login_message = Msg.Flash.LOGIN_REQUIRED
 mail = Mail()
 bootstrap = Bootstrap()
-pdfkit_config = pdfkit.configuration(
-    wkhtmltopdf=os.getenv("PDFKIT_WKHTMLTOPDF_PATH"))
 
 # Intialize google drive service
 from .google_drive import google_drive_init
@@ -71,7 +68,6 @@ def create_app(config):
     login.init_app(app)
     mail.init_app(app)
     bootstrap.init_app(app)
-    pdfkit_config.wkhtmltopdf = app.config["PDFKIT_WKHTMLTOPDF_PATH"]
 
     ################
     # Register all blueprints
